Refactored obsolete services to config
